Move Error screen to own folder + rename SuccessMessage to SuccessScreen
