+++ ws-nextjs-app/pages/[service]/send/[id]/ErrorScreen/index.stories.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import ErrorScreen from '.';
+
+const Component = () => <ErrorScreen />;
+
+export default {
+  title: 'Components/FormScreens/ErrorScreen',
+  component: Component,
+};
+
+export const Example = Component;
