--- ws-nextjs-app/pages/[service]/send/[id]/SuccessMessage/index.stories.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import SuccessMessage from './index';
-
-const Component = () => <SuccessMessage />;
-
-export default {
-  title: 'Components/SuccessMessage',
-  component: Component,
-};
-
-export const Default = Component;
