+++ ws-nextjs-app/pages/[service]/send/[id]/SuccessScreen/index.stories.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import SuccessScreen from '.';
+
+const Component = () => <SuccessScreen />;
+
+export default {
+  title: 'Components/FormScreens/SuccessScreen',
+  component: Component,
+};
+
+export const Example = Component;
