+++ ws-nextjs-app/pages/[service]/send/[id]/SuccessScreen/index.styles.ts
