+++ ws-nextjs-app/pages/[service]/send/[id]/SuccessScreen/svgs.tsx
